[rotor] add rotor-test-client to Dockerfile (#5624)
One outcome of turbinelabs/rotor#16 was that I realized it would have been super helpful to be able to dump the config rotor was serving. We have this technology already, in the form of `tbn-xds-test-client`.

I renamed it `rotor-test-client`, reworked it slightly, and added it to the Dockerfile. It's wired up now so that if you've overridden ROTOR_XDS_ADDR to be some other wacky port, `rotor-test-client` will pick that up by default. This means that in the standard case, all you have to do is run this to get a dump:

```
docker exec <rotor container id> rotor-test-client
```

fbshipit-source-id: 0623828