[rotor] add ClusterLoadAssignment (EDS) to static resources (#5975)
This is a partial fix for #5925; by allowing specification of load assignments in static-resources, we enable the ability to change the instance list for a given cluster. The practical use for this is to override one or more clusters to be localhost, so that a single proxy can be reused for all sidecars in a mesh.

I tested this by loading up all-in-one, which was set to load balance 50/50 blue/green, and adding the following static resources file, which caused it to return all blues instead by forcing the cluster EDS data to the blue instance:

```javascript
{
  "clusters": [
    {
      "name": "all-in-one-server",
      "type": "EDS",
      "edsClusterConfig": {
        "edsConfig": {
          "apiConfigSource": {
            "apiType": "GRPC",
            "grpcServices": [
              {
                "envoyGrpc": {
                  "clusterName": "tbn-xds"
                }
              }
            ],
            "refreshDelay": "30s"
          }
        },
        "serviceName": "all-in-one-server"
      },
      "connectTimeout": "10s",
      "lbPolicy": "LEAST_REQUEST",
      "lbSubsetConfig": {
        "fallbackPolicy": "ANY_ENDPOINT"
      }
    }
  ],
  "loadAssignments": [
    {
      "clusterName": "all-in-one-server",
      "endpoints": [
        {
          "lbEndpoints": [
            {
              "endpoint": {
                "address": {
                  "socketAddress": {
                    "address": "127.0.0.1",
                    "portValue": 8080
                  }
                }
              },
              "healthStatus": "HEALTHY"
            }
          ]
        }
      ]
    }
  ]
}
```

fbshipit-source-id: ead814a