[rotor] fix serverNames bug (#5974)
The problem: when more than one domain is multiplexed on a single port, _without_ SNI, serverNames are added to the FilterChainMatch, which causes envoy to refuse the connection and log (at the debug level) a complaint about no filter chain matching.

As-is the code solves this problem for a single non-SNI domain, but not for the multiplexed case.

I wrote an integration test and confirmed that it fails on master with connection refused, and passes with this patch.

In fixing it, I found myself having some cognitive dissonance with `nonSSLFilterChains` being a slice, so I reversed myself in a [prior discussion](https://github.com/turbinelabs/tbn/pull/5877#discussion_r209273548) about it, and converted it to a pointer instead. There is some mess at the bottom to merge the pointer with the `sslFilterChains` slice, but I think it's worth the code clarity.

fbshipit-source-id: 1fcbca7