[rotor] local cluster awareness for mesh routing (fixes #5925) (#5986)
This change makes it easier to use envoy/rotor sidecars in a service mesh, by making it possible to share a single proxy configuration across all clusters, while routing locally when appropriate.

#5975 added the ability to specify custom load assignments. This change adds a new environment variable to the rotor docker container, `ROTOR_LOCAL_CLUSTERS`, which is a comma-delimited list of cluster `<name>:<port>` pairs. When specified, rotor will override clusters matching each `<name>` with cluster containing only `localhost:<port>`.

We do this in the container, taking advantage of rotor primitives, rather than baking it directly into the rotor binary; the idea is that most of the time sidecar meshes will use the docker image, and if not, it's easy to explain how to do this by hand if necessary.

Also fix up all-in-one to re-use the `rotor.sh` script instead of rolling its own. This lets us test this functionality by setting `ROTOR_LOCAL_CLUSTERS=all-in-one-server:8080` when starting up all-in-one, which routes all server traffic to the blue instance, regardless of subset rules.

fbshipit-source-id: 35f8f65